Move the lexer away to a separate file
